dags/requirements.txt
yfinance
pandas
python-dotenv
SQLAlchemy==2.0.23